+++ packages/ui/src/components/molecules/SubmitButton/SubmitButton.tsx
@@ -1,0 +1,34 @@
+import { Button, ButtonProps } from '@components/atoms';
+import clsx from 'clsx';
+import { Loader2 } from 'lucide-react';
+import { forwardRef } from 'react';
+
+export interface SubmitButtonProps extends ButtonProps {
+  isLoading?: boolean;
+  loaderClassName?: string;
+}
+
+export const SubmitButton = forwardRef<HTMLButtonElement, SubmitButtonProps>(
+  (
+    { isLoading, children, className, loaderClassName, type = 'submit', ...restButtonProps },
+    ref,
+  ) => {
+    return (
+      <Button
+        type={type}
+        ref={ref}
+        className={clsx(className, 'relative', { 'pointer-events-none': isLoading })}
+        {...restButtonProps}
+      >
+        <span className={clsx({ invisible: isLoading })}>{children}</span>
+        {isLoading ? (
+          <div className="absolute top-0 left-0 flex h-full w-full items-center justify-center">
+            <Loader2
+              className={clsx('text-primary-foreground animate-spin text-sm', loaderClassName)}
+            />
+          </div>
+        ) : null}
+      </Button>
+    );
+  },
+);