+++ services/workflows-service/src/workflow/workflow.controller.internal.unit.test.ts
@@ -109,6 +109,7 @@ describe('WorkflowControllerInternal', () => {
       workflowDefinitionRepo as any,
       workflowRuntimeDataRepo,
       endUserRepo,
+      {} as any,
       businessRepo,
       entityRepo,
       {} as any,